bug: correct lp norm
Before I was computing the equivalent to the MSE rather than the L_2 distance. Now I compute the L_2 distance and I make sure that the relative version is sum( ( |x_i -x'_i|/(1+|x_i|) )^p )^(1/p). The addition by 1 is to avoid that flux values below zero bias the score in a thir region